services/llm-agent/requirements.txt
## Core
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4

## LLM
litellm==1.40.0
tiktoken==0.7.0
openai==1.35.3

## gRPC
grpcio==1.64.1
grpcio-tools==1.64.1
grpcio-health-checking==1.64.1
protobuf==5.27.1

## Redis
redis[hiredis]==5.0.7

## Resilience
pybreaker==1.2.0

## Logging
structlog==24.2.0

## Shared (local package)
# Installed via: pip install -e ../../shared/python
